+++ src/main/kotlin/dev/tailwindrainbow/intellij/adapter/intellij/highlighting/TailwindRainbowAnnotator.kt
@@ -3,6 +3,7 @@ package dev.tailwindrainbow.intellij.adapter.intellij.highlighting
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
+import com.intellij.openapi.editor.colors.EditorColorsManager
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
@@ -23,11 +24,12 @@ class TailwindRainbowAnnotator : Annotator {
         }
 
         val extension = element.virtualFile?.extension ?: return
+        val background = EditorColorsManager.getInstance().globalScheme.defaultBackground
 
         PluginComponents.highlightDocument(element.project).highlight(element.text, extension).forEach { segment ->
             holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
                 .range(TextRange(segment.start, segment.end))
-                .enforcedTextAttributes(segment.style.toTextAttributes())
+                .enforcedTextAttributes(segment.style.toTextAttributes(background))
                 .create()
         }
     }